add openvino supported

+++ deploy/openvino/include/paddlex/results.h
@@ -61,7 +61,7 @@ class DetResult : public BaseResult {
 
 class SegResult : public BaseResult {
  public:
-  Mask<int64_t> label_map;
+  Mask<int> label_map;
   Mask<float> score_map;
   void clear() {
     label_map.clear();
